Initialized layout level data using Vuex component. Will make a specific layout "page" type in butter after I get access again :) The layout data doesn't get loaded for SPA builds so will have to check on that if I want othe app to run as SPA.
Can no pass page data through the props on the extended routes - asyncData for some reason wouldn't obtain them through the payload so ... whatever, this is cleaner and works in spa or ssr.
Would really like to figure out how to use axios in the routes setup. Either way I need to automate loading of the the CMS data updates.